% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Puns from "The Last of Us"}
\format{
A data frame with 33 rows and 1 column:
\describe{
  \item{pun}{The puns from the Ellie's book}
  ...
}
}
\source{
\url{https://thelastofus.fandom.com/wiki/No_Pun_Intended:_Volume_Too}
}
\usage{
data
}
\description{
A data set that contains strings that are from the series "The Last of Us"
}
\keyword{datasets}
