% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nopunintended.R
\name{nopunintended}
\alias{nopunintended}
\title{Prints a pun.}
\usage{
nopunintended(number = NULL)
}
\arguments{
\item{number}{The number (i.e., 1-33) or an empty argument.}
}
\value{
A string returning the pun.
}
\description{
Prints a pun.
}
\examples{
nopunintended()
"What did the triangle say to the circle? You're so pointless."
nopunintended(14)
"You know what's not right? (Joel questions, \"Left?\") (Ellie answers, \"Yeah.\")"
}
